[ini] Make remote subdirs cumulative.
Configure will re-order the entries in alphabetical order, which can
cause unexpected results if SUBDIR is completely replaced. Instead, we
add to the SUBDIRS variable, so that the results are closer to what one
would expect.

Similar approach may be desirable for other brands, but is definitely
required for BEYONWIZ  / INI combination.