[vu] fix rcpositions.xml & remote.html
Clean up positions of hotspots in rcpositions.xml and shapes in
remote.html to more accurately reflect key positions.

Change the key name for the "open movie list" button from
PVR to VIDEO to match the correct code.

Add a key description entry for SUBTITLE, which was missing.

Generate shapes in remote.html that match the hotspot locations and
key positions.